pyproject.toml
[tool.poetry]
name = "hyundai_driving_experience"
version = "0.1.0"
description = ""
authors = ["Junoh Moon <mjo970625@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests-html = "^0.10.0"
beautifulsoup4 = "^4.11.1"
SQLAlchemy = "^1.4.40"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^22.6.0"
rope = "^1.3.0"
pre-commit = "^2.20.0"
neovim = "^0.3.1"
sqlalchemy2-stubs = "^0.0.2-alpha.26"
pytest-mock = "^3.8.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
